Rename namespace Console => Core. The project is no longer console application, it's library dll
