--- TelemetryExporter.Console/Widgets/Interfaces/IWidget.cs
@@ -1,9 +1,0 @@
-﻿using TelemetryExporter.Console.Models;
-
-namespace TelemetryExporter.Console.Widgets.Interfaces
-{
-    internal interface IWidget
-    {
-        Task GenerateImageAsync(SessionData sessionData, FrameData currentData);
-    }
-}
