+++ TelemetryExporter.Core/Images/radial_6.png
