+++ TelemetryExporter.Core/Images/radial_6_dial.png
