+++ TelemetryExporter.Core/Models/UIWidgetData.cs
@@ -1,0 +1,9 @@
+﻿namespace TelemetryExporter.Core.Models
+{
+    internal class UIWidgetData(int index, string name)
+    {
+        public int Index { get; set; } = index;
+
+        public string Name { get; set; } = name;
+    }
+}
