+++ TelemetryExporter.Core/Widgets/Interfaces/IWidget.cs
@@ -1,0 +1,9 @@
+﻿using TelemetryExporter.Core.Models;
+
+namespace TelemetryExporter.Core.Widgets.Interfaces
+{
+    internal interface IWidget
+    {
+        Task GenerateImageAsync(SessionData sessionData, FrameData currentData);
+    }
+}
